* Chaging hardware tests so that tests marked with `@pytest.mark.hardware` are skipped unless `--hardware-test` is given on command line. * Moving all test files into `pocs/tests` for now as it makes the conftest cleaner.
