+++ pocs/tests/test_boards.py
@@ -1,0 +1,17 @@
+import pytest
+
+from peas.sensors import ArduinoSerialMonitor
+
+
+@pytest.fixture(scope='module')
+def monitor():
+    return ArduinoSerialMonitor(auto_detect=True)
+
+
+def test_create(monitor):
+    assert monitor is not None
+
+
+@pytest.mark.hardware
+def test_has_readers(monitor):
+    assert len(monitor.serial_readers) > 0
